feat/Visual Code Styles
