+++ src/Pages/dashboard/Editor/codeScreen/CodePanel/CodePanel.module.css
@@ -1,6 +1,7 @@
 .codePanel {
   background-color: #fafafa;
-  width: 100vw;
+  width: 1250px;
+  height: 660px;
   top: 3.7rem;
   bottom: 0;
   left: 31.4vh;
@@ -16,15 +17,14 @@
   display: none;
 }
 
-.containerOptions button {
-  /* border: 2px solid blue; */
-  cursor: pointer;
-  font-weight: 600;
+.containerOptionesCodePanel{
+  height: 200px;
 }
 
 .languageOptions {
   display: flex;
   justify-content: space-between;
+  width: fit-content;
 }
 
 
@@ -74,9 +74,6 @@
   height: 80vh;
   border: 0;
   outline: 0;
-  margin: auto;
-  padding: auto;
-  margin-top: 15px;
   overflow: scroll;
 }
 
@@ -108,32 +105,6 @@
 
 .background {
   background-color: #fafafa !important;
-}
-
-/* Terminal */
-.codeTerminal {
-  border-top: 1px solid #36363660;
-  align-self: flex-end;
-  height: 30vh;
-  width: 100%;
-  margin-top: -20vh;
-  z-index: 10;
-}
-
-.consoleMenu {
-  display: flex;
-  gap: 10px;
-}
-
-.terminalActive {
-  padding-bottom: 5px;
-  border-bottom: 2px solid #36363660;
-}
-
-.terminalOption {
-  padding-bottom: 5px;
-  border-bottom: 2px solid transparent;
-  color: #36363660;
 }
 
 .codeContentShow {
